+++ lib/ibanity/api/consent.rb
@@ -1,0 +1,2 @@
+require_relative "consent/consent"
+require_relative "consent/processing_operation"