+++ lib/ibanity/api/sandbox/financial_institution_account.rb
@@ -1,40 +1,25 @@
 module Ibanity
   module Sandbox
+    # <b>DEPRECATED:</b> Please use <tt>Ibanity::Xs2a::Sandbox::FinancialInstitutionAccount</tt> instead.
     class FinancialInstitutionAccount < Ibanity::BaseResource
       def self.create(financial_institution_user_id:, financial_institution_id:, idempotency_key: nil, **attributes)
-        path = Ibanity.sandbox_api_schema["financialInstitution"]["financialInstitutionAccounts"]
-          .gsub("{financialInstitutionId}", financial_institution_id)
-          .gsub("{financialInstitutionUserId}", financial_institution_user_id)
-          .gsub("{financialInstitutionAccountId}", "")
-        uri = Ibanity.client.build_uri(path)
-        create_by_uri(uri: uri, resource_type: "financialInstitutionAccount", attributes: attributes, idempotency_key: idempotency_key)
+        warn "[DEPRECATION] `Ibanity::Sandbox::FinancialInstitutionAccount.create` is deprecated. Please use Ibanity::Xs2a::Sandbox::FinancialInstitutionAccount.create instead"
+        Ibanity::Xs2a::Sandbox::FinancialInstitutionAccount.create(financial_institution_user_id: financial_institution_user_id, financial_institution_id: financial_institution_id, idempotency_key: idempotency_key, attributes)
       end
 
       def self.list(financial_institution_id:, financial_institution_user_id:, **query_params)
-        path = Ibanity.sandbox_api_schema["financialInstitution"]["financialInstitutionAccounts"]
-          .gsub("{financialInstitutionId}", financial_institution_id)
-          .gsub("{financialInstitutionUserId}", financial_institution_user_id)
-          .gsub("{financialInstitutionAccountId}", "")
-        uri = Ibanity.client.build_uri(path)
-        list_by_uri(uri: uri, query_params: query_params)
+        warn "[DEPRECATION] `Ibanity::Sandbox::FinancialInstitutionAccount.list` is deprecated. Please use Ibanity::Xs2a::Sandbox::FinancialInstitutionAccount.list instead"
+        Ibanity::Xs2a::Sandbox::FinancialInstitutionAccount.list(financial_institution_id: financial_institution_id, financial_institution_user_id: financial_institution_user_id, query_params)
       end
 
       def self.find(id:, financial_institution_user_id:, financial_institution_id:)
-        path = Ibanity.sandbox_api_schema["financialInstitution"]["financialInstitutionAccounts"]
-          .gsub("{financialInstitutionId}", financial_institution_id)
-          .gsub("{financialInstitutionUserId}", financial_institution_user_id)
-          .gsub("{financialInstitutionAccountId}", id)
-        uri = Ibanity.client.build_uri(path)
-        find_by_uri(uri: uri)
+        warn "[DEPRECATION] `Ibanity::Sandbox::FinancialInstitutionAccount.find` is deprecated. Please use Ibanity::Xs2a::Sandbox::FinancialInstitutionAccount.find instead"
+        Ibanity::Xs2a::Sandbox::FinancialInstitutionAccount.find(id: id, financial_institution_user_id: financial_institution_user_id, financial_institution_id: financial_institution_id)
       end
 
       def self.delete(id:, financial_institution_user_id:, financial_institution_id:)
-        path = Ibanity.sandbox_api_schema["financialInstitution"]["financialInstitutionAccounts"]
-          .gsub("{financialInstitutionId}", financial_institution_id)
-          .gsub("{financialInstitutionUserId}", financial_institution_user_id)
-          .gsub("{financialInstitutionAccountId}", id)
-        uri = Ibanity.client.build_uri(path)
-        destroy_by_uri(uri: uri)
+        warn "[DEPRECATION] `Ibanity::Sandbox::FinancialInstitutionAccount.delete` is deprecated. Please use Ibanity::Xs2a::Sandbox::FinancialInstitutionAccount.delete instead"
+        Ibanity::Xs2a::Sandbox::FinancialInstitutionAccount.delete(id: id, financial_institution_user_id: financial_institution_user_id, financial_institution_id: financial_institution_id)
       end
     end
   end