+++ lib/ibanity/api/sandbox/financial_institution_holding.rb
@@ -1,44 +1,25 @@
 module Ibanity
   module Sandbox
+    # <b>DEPRECATED:</b> Please use <tt>Ibanity::Xs2a::Sandbox::FinancialInstitutionHolding</tt> instead.
     class FinancialInstitutionHolding < Ibanity::BaseResource
       def self.create(financial_institution_user_id:, financial_institution_id:, financial_institution_account_id:, idempotency_key: nil, **attributes)
-        path = Ibanity.sandbox_api_schema["financialInstitution"]["financialInstitutionAccount"]["financialInstitutionHoldings"]
-          .gsub("{financialInstitutionId}", financial_institution_id)
-          .gsub("{financialInstitutionUserId}", financial_institution_user_id)
-          .gsub("{financialInstitutionAccountId}", financial_institution_account_id)
-          .gsub("{financialInstitutionHoldingId}", "")
-        uri = Ibanity.client.build_uri(path)
-        create_by_uri(uri: uri, resource_type: "financialInstitutionHolding", attributes: attributes, idempotency_key: idempotency_key)
+        warn "[DEPRECATION] `Ibanity::Sandbox::FinancialInstitutionHolding.create` is deprecated. Please use Ibanity::Xs2a::Sandbox::FinancialInstitutionHolding.create instead"
+        Ibanity::Xs2a::Sandbox::FinancialInstitutionHolding.create(financial_institution_user_id: financial_institution_user_id, financial_institution_id: financial_institution_id, financial_institution_account_id: financial_institution_account_id, idempotency_key: idempotency_key, attributes)
       end
 
       def self.list(financial_institution_user_id:, financial_institution_id:, financial_institution_account_id:, **query_params)
-        path = Ibanity.sandbox_api_schema["financialInstitution"]["financialInstitutionAccount"]["financialInstitutionHoldings"]
-          .gsub("{financialInstitutionId}", financial_institution_id)
-          .gsub("{financialInstitutionUserId}", financial_institution_user_id)
-          .gsub("{financialInstitutionAccountId}", financial_institution_account_id)
-          .gsub("{financialInstitutionHoldingId}", "")
-        uri = Ibanity.client.build_uri(path)
-        list_by_uri(uri: uri, query_params: query_params)
+        warn "[DEPRECATION] `Ibanity::Sandbox::FinancialInstitutionHolding.list` is deprecated. Please use Ibanity::Xs2a::Sandbox::FinancialInstitutionHolding.list instead"
+        Ibanity::Xs2a::Sandbox::FinancialInstitutionHolding.list(financial_institution_user_id: financial_institution_user_id, financial_institution_id: financial_institution_id, financial_institution_account_id: financial_institution_account_id, query_params)
       end
 
       def self.find(id:, financial_institution_user_id:, financial_institution_id:, financial_institution_account_id:)
-        path = Ibanity.sandbox_api_schema["financialInstitution"]["financialInstitutionAccount"]["financialInstitutionHoldings"]
-          .gsub("{financialInstitutionId}", financial_institution_id)
-          .gsub("{financialInstitutionUserId}", financial_institution_user_id)
-          .gsub("{financialInstitutionAccountId}", financial_institution_account_id)
-          .gsub("{financialInstitutionHoldingId}", id)
-        uri = Ibanity.client.build_uri(path)
-        find_by_uri(uri: uri)
+        warn "[DEPRECATION] `Ibanity::Sandbox::FinancialInstitutionHolding.find` is deprecated. Please use Ibanity::Xs2a::Sandbox::FinancialInstitutionHolding.find instead"
+        Ibanity::Xs2a::Sandbox::FinancialInstitutionHolding.find(id: id, financial_institution_user_id: financial_institution_user_id, financial_institution_id: financial_institution_id, financial_institution_account_id: financial_institution_account_id)
       end
 
       def self.delete(id:, financial_institution_user_id:, financial_institution_id:, financial_institution_account_id:)
-        path = Ibanity.sandbox_api_schema["financialInstitution"]["financialInstitutionAccount"]["financialInstitutionHoldings"]
-          .gsub("{financialInstitutionId}", financial_institution_id)
-          .gsub("{financialInstitutionUserId}", financial_institution_user_id)
-          .gsub("{financialInstitutionAccountId}", financial_institution_account_id)
-          .gsub("{financialInstitutionHoldingId}", id)
-        uri = Ibanity.client.build_uri(path)
-        destroy_by_uri(uri: uri)
+        warn "[DEPRECATION] `Ibanity::Sandbox::FinancialInstitutionHolding.delete` is deprecated. Please use Ibanity::Xs2a::Sandbox::FinancialInstitutionHolding.delete instead"
+        Ibanity::Xs2a::Sandbox::FinancialInstitutionHolding.find(id: id, financial_institution_user_id: financial_institution_user_id, financial_institution_id: financial_institution_id, financial_institution_account_id: financial_institution_account_id)
       end
     end
   end