+++ lib/ibanity/api/sandbox/financial_institution_transaction.rb
@@ -1,54 +1,30 @@
 module Ibanity
   module Sandbox
+    # <b>DEPRECATED:</b> Please use <tt>Ibanity::Xs2a::Sandbox::FinancialInstitutionTransaction</tt> instead.
     class FinancialInstitutionTransaction < Ibanity::BaseResource
       def self.create(financial_institution_user_id:, financial_institution_id:, financial_institution_account_id:, idempotency_key: nil, **attributes)
-        path = Ibanity.sandbox_api_schema["financialInstitution"]["financialInstitutionAccount"]["financialInstitutionTransactions"]
-          .gsub("{financialInstitutionId}", financial_institution_id)
-          .gsub("{financialInstitutionUserId}", financial_institution_user_id)
-          .gsub("{financialInstitutionAccountId}", financial_institution_account_id)
-          .gsub("{financialInstitutionTransactionId}", "")
-        uri = Ibanity.client.build_uri(path)
-        create_by_uri(uri: uri, resource_type: "financialInstitutionTransaction", attributes: attributes, idempotency_key: idempotency_key)
+        warn "[DEPRECATION] `Ibanity::Sandbox::FinancialInstitutionTransaction.create` is deprecated. Please use Ibanity::Xs2a::Sandbox::FinancialInstitutionTransaction.create instead"
+        Ibanity::Xs2a::Sandbox::FinancialInstitutionTransaction.create(financial_institution_user_id: financial_institution_user_id, financial_institution_id: financial_institution_id, financial_institution_account_id: financial_institution_account_id, idempotency_key: idempotency_key, attributes)
       end
 
       def self.list(financial_institution_user_id:, financial_institution_id:, financial_institution_account_id:, **query_params)
-        path = Ibanity.sandbox_api_schema["financialInstitution"]["financialInstitutionAccount"]["financialInstitutionTransactions"]
-          .gsub("{financialInstitutionId}", financial_institution_id)
-          .gsub("{financialInstitutionUserId}", financial_institution_user_id)
-          .gsub("{financialInstitutionAccountId}", financial_institution_account_id)
-          .gsub("{financialInstitutionTransactionId}", "")
-        uri = Ibanity.client.build_uri(path)
-        list_by_uri(uri: uri, query_params: query_params)
+        warn "[DEPRECATION] `Ibanity::Sandbox::FinancialInstitutionTransaction.list` is deprecated. Please use Ibanity::Xs2a::Sandbox::FinancialInstitutionTransaction.list instead"
+        Ibanity::Xs2a::Sandbox::FinancialInstitutionTransaction.list(financial_institution_user_id: financial_institution_user_id, financial_institution_id: financial_institution_id, financial_institution_account_id: financial_institution_account_id, query_params)
       end
 
       def self.find(id:, financial_institution_user_id:, financial_institution_id:, financial_institution_account_id:)
-        path = Ibanity.sandbox_api_schema["financialInstitution"]["financialInstitutionAccount"]["financialInstitutionTransactions"]
-          .gsub("{financialInstitutionId}", financial_institution_id)
-          .gsub("{financialInstitutionUserId}", financial_institution_user_id)
-          .gsub("{financialInstitutionAccountId}", financial_institution_account_id)
-          .gsub("{financialInstitutionTransactionId}", id)
-        uri = Ibanity.client.build_uri(path)
-        find_by_uri(uri: uri)
+        warn "[DEPRECATION] `Ibanity::Sandbox::FinancialInstitutionTransaction.find` is deprecated. Please use Ibanity::Xs2a::Sandbox::FinancialInstitutionTransaction.find instead"
+        Ibanity::Xs2a::Sandbox::FinancialInstitutionTransaction.find(id: id, financial_institution_user_id: financial_institution_user_id, financial_institution_id: financial_institution_id, financial_institution_account_id: financial_institution_account_id)
       end
 
       def self.update(id:, financial_institution_user_id:, financial_institution_id:, financial_institution_account_id:, idempotency_key: nil, **attributes)
-        path = Ibanity.sandbox_api_schema["financialInstitution"]["financialInstitutionAccount"]["financialInstitutionTransactions"]
-          .gsub("{financialInstitutionId}", financial_institution_id)
-          .gsub("{financialInstitutionUserId}", financial_institution_user_id)
-          .gsub("{financialInstitutionAccountId}", financial_institution_account_id)
-          .gsub("{financialInstitutionTransactionId}", id)
-        uri = Ibanity.client.build_uri(path)
-        update_by_uri(uri: uri, resource_type: "financialInstitutionTransaction", attributes: attributes, idempotency_key: idempotency_key)
+        warn "[DEPRECATION] `Ibanity::Sandbox::FinancialInstitutionTransaction.create` is deprecated. Please use Ibanity::Xs2a::Sandbox::FinancialInstitutionTransaction.create instead"
+        Ibanity::Xs2a::Sandbox::FinancialInstitutionTransaction.create(id: id, financial_institution_user_id: financial_institution_user_id, financial_institution_id: financial_institution_id, financial_institution_account_id: financial_institution_account_id, idempotency_key: idempotency_key, attributes)
       end
 
       def self.delete(id:, financial_institution_user_id:, financial_institution_id:, financial_institution_account_id:)
-        path = Ibanity.sandbox_api_schema["financialInstitution"]["financialInstitutionAccount"]["financialInstitutionTransactions"]
-          .gsub("{financialInstitutionId}", financial_institution_id)
-          .gsub("{financialInstitutionUserId}", financial_institution_user_id)
-          .gsub("{financialInstitutionAccountId}", financial_institution_account_id)
-          .gsub("{financialInstitutionTransactionId}", id)
-        uri = Ibanity.client.build_uri(path)
-        destroy_by_uri(uri: uri)
+        warn "[DEPRECATION] `Ibanity::Sandbox::FinancialInstitutionTransaction.delete` is deprecated. Please use Ibanity::Xs2a::Sandbox::FinancialInstitutionTransaction.delete instead"
+        Ibanity::Xs2a::Sandbox::FinancialInstitutionTransaction.find(id: id, financial_institution_user_id: financial_institution_user_id, financial_institution_id: financial_institution_id, financial_institution_account_id: financial_institution_account_id)
       end
     end
   end