+++ lib/ibanity/api/xs2a/transaction.rb
@@ -6,7 +6,7 @@ module Ibanity
           Ibanity.xs2a_api_schema["customer"]["synchronization"]["updatedTransactions"]
           .sub("{synchronizationId}", synchronization_id)
         else
-          Ibanity.xs2a_api_schema["customer"]["financialInstitution"]["transactions"]
+          Ibanity.xs2a_api_schema["customer"]["financialInstitution"]["account"]["transactions"]
             .sub("{financialInstitutionId}", financial_institution_id)
             .sub("{accountId}", account_id)
             .sub("{transactionId}", "")
@@ -15,7 +15,7 @@ module Ibanity
       end
 
       def self.find(id:, financial_institution_id:, account_id:, customer_access_token:)
-        uri = Ibanity.xs2a_api_schema["customer"]["financialInstitution"]["transactions"]
+        uri = Ibanity.xs2a_api_schema["customer"]["financialInstitution"]["account"]["transactions"]
           .sub("{financialInstitutionId}", financial_institution_id)
           .sub("{accountId}", account_id)
           .sub("{transactionId}", id)